Type disruption sector state

+++ src/yt_game.h
@@ -37,7 +37,7 @@ struct yt_player {
 
 bool yt_game_load_startup_configuration(struct yt_game *game,
 	const char *path, bool local_mode, struct yt_player_cache *player_cache,
-	float disruption_sectors[2], float *local_screen,
+	int disruption_sectors[2], float *local_screen,
 	struct yt_error *error);
 
 struct yt_sector {
